Cleanup shading stuff
The GetShade/shadetable stuff was handled in a weird way which led to a lot of repeated code. Now GetShade returns a pointer to the shadetable, and also takes a flags argument so it knows when to use the original colors (if FL_FULLBRIGHT is set). Also, "curshades" was such a terrible name for the shade pointer, so now it's just called "shade".